Minor refactor of RulesEngine package
